+++ examples/base_cc_gwlb/main.tf
@@ -43,6 +43,7 @@ EOF
   }
 }
 
+
 ## Create the user_data file
 locals {
   userdata = <<USERDATA
@@ -64,6 +65,7 @@ resource "local_file" "user-data-file" {
 data "aws_availability_zones" "available" {
   state = "available"
 }
+
 
 # Create a new VPC
 resource "aws_vpc" "vpc1" {
@@ -111,7 +113,7 @@ resource "aws_subnet" "privsubnet" {
 }
 
 
-# Create a public Route Table towards IGW.
+# Create a public Route Table towards IGW
 resource "aws_route_table" "routetablepublic1" {
   vpc_id = aws_vpc.vpc1.id
 
@@ -125,11 +127,12 @@ resource "aws_route_table" "routetablepublic1" {
   )
 }
 
+
 # Create equal number of Route Table associations to how many Public subnets exist. 
 resource "aws_route_table_association" "routetablepublic1" {
-  count = length(aws_subnet.pubsubnet.*.id)
-  subnet_id      = aws_subnet.pubsubnet.*.id[count.index]
-  route_table_id = aws_route_table.routetablepublic1.id
+  count           = length(aws_subnet.pubsubnet.*.id)
+  subnet_id       = aws_subnet.pubsubnet.*.id[count.index]
+  route_table_id  = aws_route_table.routetablepublic1.id
 }
 
 
@@ -158,10 +161,9 @@ resource "aws_nat_gateway" "ngw" {
 }
 
 
-
 # 2. Create Bastion Host
 module "bastion" {
-  source        = "../../modules/terraform-zsbastion-aws"
+  source        = "../../modules/terraform-zscc-bastion-aws"
   name_prefix   = var.name_prefix
   resource_tag  = random_string.suffix.result
   global_tags   = local.global_tags
@@ -175,20 +177,19 @@ module "bastion" {
 # 3. Create Workload
 # Create Workloads
 module "workload" {
-  workload_count = var.workload_count
-  source       = "../../modules/terraform-zsworkload-aws"
-  name_prefix  = "${var.name_prefix}-workload"
-  resource_tag = random_string.suffix.result
-  global_tags  = local.global_tags
-  vpc          = aws_vpc.vpc1.id
-  subnet       = aws_subnet.privsubnet.*.id
-  instance_key = aws_key_pair.deployer.key_name
+  workload_count  = var.workload_count
+  source          = "../../modules/terraform-zscc-workload-aws"
+  name_prefix     = "${var.name_prefix}-workload"
+  resource_tag    = random_string.suffix.result
+  global_tags     = local.global_tags
+  vpc             = aws_vpc.vpc1.id
+  subnet          = aws_subnet.privsubnet.*.id
+  instance_key    = aws_key_pair.deployer.key_name
 }
 
 
-
 # 4. Create CC network, routing, and appliance
-# Create subnet for CC network in X availability zones per az_count variable
+# Create subnet for CC network
 resource "aws_subnet" "cc-subnet" {
   count = var.az_count
 
@@ -210,7 +211,7 @@ resource "aws_route_table" "cc-rt" {
     cidr_block     = "0.0.0.0/0"
     nat_gateway_id = element(aws_nat_gateway.ngw.*.id, count.index)
   }
-
+  
   tags = merge(local.global_tags,
         { Name = "${var.name_prefix}-cc-rt-ngw-${count.index + 1}-${random_string.suffix.result}" }
   )
@@ -235,22 +236,49 @@ EOF
   }
 }
 
+
+
 # Create X CC VMs per cc_count which will span equally across designated availability zones per az_count
 # E.g. cc_count set to 4 and az_count set to 2 will create 2x CCs in AZ1 and 2x CCs in AZ2
 module "cc-vm" {
-  source              = "../../modules/terraform-zscc-aws"
-  cc_count            = var.cc_count
+  source                    = "../../modules/terraform-zscc-ccvm-aws"
+  cc_count                  = var.cc_count
+  name_prefix               = var.name_prefix
+  resource_tag              = random_string.suffix.result
+  global_tags               = local.global_tags
+  vpc                       = aws_vpc.vpc1.id
+  mgmt_subnet_id            = aws_subnet.cc-subnet.*.id
+  service_subnet_id         = aws_subnet.cc-subnet.*.id
+  instance_key              = aws_key_pair.deployer.key_name
+  user_data                 = local.userdata
+  ccvm_instance_type        = var.ccvm_instance_type
+  cc_instance_size          = var.cc_instance_size
+  iam_instance_profile      = module.cc-iam.iam_instance_profile_id
+  mgmt_security_group_id    = module.cc-sg.mgmt_security_group_id
+  service_security_group_id = module.cc-sg.service_security_group_id
+}
+
+
+# Create IAM Policy, Roles, and Instance Profiles to be assigned to CC appliances. Default behavior will create 1 of each resource per CC VM. Set variable reuse_iam to true
+# if you would like a single IAM profile created and assigned to ALL Cloud Connectors
+module "cc-iam" {
+  source              = "../../modules/terraform-zscc-iam-aws"
+  iam_count           = var.reuse_iam == false ? var.cc_count : 1
   name_prefix         = var.name_prefix
   resource_tag        = random_string.suffix.result
   global_tags         = local.global_tags
-  vpc                 = aws_vpc.vpc1.id
-  mgmt_subnet_id      = aws_subnet.cc-subnet.*.id
-  service_subnet_id   = aws_subnet.cc-subnet.*.id
-  instance_key        = aws_key_pair.deployer.key_name
-  user_data           = local.userdata
-  ccvm_instance_type  = var.ccvm_instance_type
-  cc_instance_size    = var.cc_instance_size
   cc_callhome_enabled = var.cc_callhome_enabled
+}
+
+# Create Security Group and rules to be assigned to CC mgmt and and service interface(s). Default behavior will create 1 of each resource per CC VM. Set variable reuse_security_group
+# to true if you would like a single security group created and assigned to ALL Cloud Connectors
+module "cc-sg" {
+  source        = "../../modules/terraform-zscc-sg-aws"
+  sg_count      = var.reuse_security_group == false ? var.cc_count : 1
+  name_prefix   = var.name_prefix
+  resource_tag  = random_string.suffix.result
+  global_tags   = local.global_tags
+  vpc           = aws_vpc.vpc1.id
 }
 
 
@@ -258,7 +286,7 @@ module "cc-vm" {
 # 5. Create GWLB in all CC subnets. Create Target Group and attach primary service IP from all created Cloud
 #    Connectors as registered targets.
 module "gwlb" {
-  source                    = "../../modules/terraform-zsgwlb-aws"
+  source                    = "../../modules/terraform-zscc-gwlb-aws"
   name_prefix               = var.name_prefix
   resource_tag              = random_string.suffix.result
   global_tags               = local.global_tags
@@ -278,13 +306,13 @@ module "gwlb" {
 
 # 6. Create Endpoint Service associated with GWLB and 1x GWLB Endpoint per CC subnet
 module "gwlb-endpoint" {
-  source                  = "../../modules/terraform-zsgwlbendpoint-aws"
-  name_prefix             = var.name_prefix
-  resource_tag            = random_string.suffix.result
-  global_tags             = local.global_tags
-  vpc                     = aws_vpc.vpc1.id
-  cc_subnet_ids           = aws_subnet.cc-subnet.*.id
-  gwlb_arn                = module.gwlb.gwlb_arn
+  source        = "../../modules/terraform-zscc-gwlbendpoint-aws"
+  name_prefix   = var.name_prefix
+  resource_tag  = random_string.suffix.result
+  global_tags   = local.global_tags
+  vpc           = aws_vpc.vpc1.id
+  cc_subnet_ids = aws_subnet.cc-subnet.*.id
+  gwlb_arn      = module.gwlb.gwlb_arn
 }
 
 
@@ -312,66 +340,3 @@ resource "aws_route_table_association" "private-rt-asssociation" {
   subnet_id      = aws_subnet.privsubnet.*.id[count.index]
   route_table_id = aws_route_table.routetableprivate.*.id[count.index]
 }
-
-
-
-
-############################################################################################################################################
-####### Legacy code for reference if customer desires to break cloud connector mgmt and service interfaces out into separate subnets #######
-############################################################################################################################################
-
-# create new subnet for CC mgmt n/w
-#resource "aws_subnet" "cc-mgmt-subnet" {
-#  count = 2
-#
-#  availability_zone = data.aws_availability_zones.available.names[count.index]
-#  cidr_block        = cidrsubnet(aws_vpc.vpc1.cidr_block, 12, (count.index * 16) + 3936)
-#  vpc_id            = aws_vpc.vpc1.id
-#
-#  tags = map(
-#    "Name", "${var.name_prefix}-vpc1-ec-mgmt-subnet-${count.index + 1}-${random_string.suffix.result}",
-#    "zs-edge-connector-cluster/${var.name_prefix}-cluster-${random_string.suffix.result}", "shared",
-#  )
-#}
-
-#CC Mgmt/Service NATGW Route Table
-#resource "aws_route_table" "routetable-cc-mgmt-and-service" {
-#  count  = 1
-#  vpc_id = aws_vpc.vpc1.id
-#  route {
-#    cidr_block     = "0.0.0.0/0"
-#    nat_gateway_id = aws_nat_gateway.ngw1.id
-#  }
-#
-#  tags = map(
-#    "Name", "${var.name_prefix}-natgw-cc-mgmt-svc-rt-${count.index + 1}-${random_string.suffix.result}",
-#    "zs-edge-connector-cluster/${var.name_prefix}-cluster-${random_string.suffix.result}", "shared",
-#  )
-#}
-
-#CC Mgmt subnet NATGW Route Table Association
-#resource "aws_route_table_association" "routetable-cc-mgmt" {
-#  count          = 2
-#  subnet_id      = aws_subnet.cc-mgmt-subnet.*.id[count.index]
-#  route_table_id = aws_route_table.routetable-cc-mgmt-and-service.*.id[0]
-#}
-
-# create new subnet for CC service n/w
-#resource "aws_subnet" "cc-service-subnet" {
-#  count             = 2
-#  availability_zone = data.aws_availability_zones.available.names[count.index]
-#  cidr_block        = cidrsubnet(aws_vpc.vpc1.cidr_block, 12, (count.index * 16) + 4000)
-#  vpc_id            = aws_vpc.vpc1.id
-#
-#  tags = map(
-#    "Name", "${var.name_prefix}-ec-service-subnet-${count.index + 1}-${random_string.suffix.result}",
-#    "zs-edge-connector-cluster/${var.name_prefix}-cluster-${random_string.suffix.result}", "shared",
-#  )
-#}
-
-#EC Service subnet NATGW Route Table Association
-#resource "aws_route_table_association" "routetable-cc-service" {
-#  count          = 2
-#  subnet_id      = aws_subnet.cc-service-subnet.*.id[count.index]
-#  route_table_id = aws_route_table.routetable-cc-mgmt-and-service.*.id[0]
-#}