+++ modules/terraform-zscc-ccvm-aws/versions.tf
@@ -4,10 +4,6 @@ terraform {
       source  = "hashicorp/aws"
       version = "~> 4.7.0"
     }
-    random = {
-      source  = "hashicorp/random"
-      version = "~> 3.3.0"
-    }
     local = {
       source = "hashicorp/local"
     }