+++ modules/terraform-zscc-workload-aws/versions.tf
@@ -1,7 +1,8 @@
 terraform {
   required_providers {
     aws = {
-      source = "hashicorp/aws"
+      source  = "hashicorp/aws"
+      version = "~> 4.7.0"
     }
   }
   required_version = ">= 0.13"